Fix race condition in setTimeout test.
Backburner will erroneously fire only the first timer in the event that no
timers are eligible to fire.  This will actually never happen in real life, but
happens in the tests because a side effect of ensuring our two timers are
scheduled for the same time is that we will set our timer to wait 0ms.

The race condition will only trigger if the time between the restoring
Date.prototype.valueOf and the timer ticking is less than 1 millisecond, which
is why it never happened in Chrome, but consistently happened in phantom.