[fixes #51] reduce search space for pushUnique on targets with GUID_KEYS.
Although this does increase C, it does greatly improve the most common
pushUnique use-cases. Typcially, pushUniques are distributed across many
targets most often in ember these targets have a unique identifier (GUID
KEY). Utilizing this key, we create a unique Queue per target. This
nicely reducing the common search space.

So where previously a pushUnique would have had to traverse the global
queue it now must only search a queue specific to it's target.

This PR is to be paired with an ember PR that re-introduces GUIDs on
views, and SimpleHandlebarsView. These guids are extremely cheap, as
they are simply numerically increasing and do not require non
enumerability.