reduce search space for pushUnique on targets with GUID_KEYS.
Although this does increase C, it does greatly improve overall throughput as each target (with key) now has its own local index for uniqueness. So where previously a pushUnique would have had to traverse the entire queue for N

This PR is to be paired with an ember PR that re-introduces GUIDs on views, and SimpleHandlebarsView. These guids are extremely cheap, as they are simply numerically increasing and do not require non enumerability.