S3 publishing
In order for the publishing to work the following variables have to be
set

S3_BUCKET_NAME
S3_SECRET_ACCESS_KEY
S3_ACCESS_KEY_ID
inside the .travis.yml. I am unable to set this at the moment, so this
is something that should be kept in mind during merge.