Fix #135 and Cleanup Timers
Cleanup timers to make clear what is going on. Simplify based on the fact we are inserting into the _timers in sorted order.  Remove multiple sources of truth.  Simplify stale timer handling (should we move this check in other spots like run)?

Remove duplicate test of hanging issue, the other one with the spin is more reliable.