Make `now` configurable so time can freeze during backburner run loop
This allow us to use fake timers, such as `sinon.useFakeTimers`, to
check the state of the app during the run loop.

For example: if a popup is supposed to be triggered in X ms from now and
it will show to the user during Y ms, then this will allow us to use
fake timers to pause between X and Y and check the state of the Ember
app at that exact time.

This is related with #263, #179 and #166.

If someone needs to have the backburner clock separate from the global
stubbed clock (as explained in #179), then `_platform.now` config should
be used to pass the native `Date.now` function.