Add support for Chrome's Async Stack Tagging API
https://developer.chrome.com/blog/devtools-modern-web-debugging/#linked-stack-traces

This enables the Chrome developer tools to link the stack traces of the original event scheduling and the eventual execution on the runloop. This is available in Chrome 106 and above.

To protect production performance, this is disabled-by-default, but can be enabled by setting `Backburner.ASYNC_STACKS = true`. Applications/frameworks could choose to enable this by default in development modes.